sync the latest code from remote dev branch

+++ update.sh
@@ -1,5 +1,12 @@
 #!/bin/bash -e
 
+if [ -d data-backup ]; then
+   echo "ERROR: Backup directory exists. May be previous restoring was failed?"
+   echo "1. Save 'data-backup' and 'data' dirs to safe location to make possibility to restore config later."
+   echo "2. Manually delete 'data-backup' dir and try again."
+   exit 1
+fi
+
 echo "Stopping Zigbee2MQTT..."
 sudo systemctl stop zigbee2mqtt
